feat: make setup
